Address feedback from Xuezhao
Features: rebuild
Allow-unstable-test: true
Skip-list: test_ec_degrade:DAOS-15843
