+++ src/rebuild/scan.c
@@ -22,6 +22,7 @@
 #include <daos_srv/daos_mgmt_srv.h>
 #include <daos_srv/daos_engine.h>
 #include <daos_srv/rebuild.h>
+#include <daos_srv/object.h>
 #include <daos_srv/vos.h>
 #include <daos_srv/dtx_srv.h>
 #include "rpc.h"
